@@ -1,6 +1,6 @@
 {
     "name": "blog.ptcookie.dev",
-    "version": "1.0.7",
+    "version": "1.1.0",
     "description": "DevLog of PtCookie",
     "license": "MIT",
     "author": {
@@ -26,11 +26,12 @@
         "gray-matter": "^4.0.3",
         "highlight.js": "^11.8.0",
         "js-yaml": "^4.1.0",
-        "next": "^13.4.4",
+        "next": "^13.4.6",
+        "next-intl": "^2.14.6",
         "next-mdx-remote": "^4.4.1",
         "postcss": "^8.4.24",
         "postcss-flexbugs-fixes": "^5.0.2",
-        "postcss-preset-env": "^8.4.2",
+        "postcss-preset-env": "^8.5.0",
         "react": "^18.2.0",
         "react-dom": "^18.2.0",
         "react-icons": "^4.9.0",
@@ -40,22 +41,22 @@
         "yaml-loader": "^0.8.0"
     },
     "devDependencies": {
-        "@playwright/test": "^1.34.3",
+        "@playwright/test": "^1.35.1",
         "@types/chai": "^4.3.5",
         "@types/js-yaml": "^4.0.5",
         "@types/mocha": "^10.0.1",
-        "@types/node": "^20.2.5",
-        "@types/react": "^18.2.9",
-        "@types/react-dom": "^18.2.4",
+        "@types/node": "^20.3.1",
+        "@types/react": "^18.2.12",
+        "@types/react-dom": "^18.2.5",
         "chai": "^4.3.7",
-        "eslint": "^8.42.0",
-        "eslint-config-next": "^13.4.4",
+        "eslint": "^8.43.0",
+        "eslint-config-next": "^13.4.6",
         "eslint-config-prettier": "^8.8.0",
         "husky": "^8.0.3",
         "lint-staged": "^13.2.2",
         "mocha": "^10.2.0",
-        "prettier": "2.8.0",
-        "stylelint": "^15.7.0",
+        "prettier": "2.8.8",
+        "stylelint": "^15.8.0",
         "stylelint-config-standard": "^33.0.0",
         "ts-node": "^10.9.1"
     },
