@@ -1,5 +1,8 @@
-import { useRouter } from 'next/router';
+import { usePathname } from 'next/navigation';
+import { usePathname as useBasePath } from 'next-intl/client';
+import Link from 'next-intl/link';
 import { RxGlobe } from 'react-icons/rx';
+import { i18n } from '@/lib/i18n';
 
 import styles from './LocaleSelector.module.css';
 
@@ -8,22 +11,20 @@ type Props = {
 };
 
 export default function LocaleSelector({ active }: Props) {
-  const router = useRouter();
-  const { pathname, asPath, query, locales, locale: currentLocale } = router;
+  const pathname = usePathname();
+  const basePath = useBasePath();
+
+  const currentLocale = pathname.split('/')?.[1];
+  const { locales } = i18n;
 
   return (
     <ul className={styles.locale} data-active={active}>
       <RxGlobe />
       {locales?.map((locale) => (
-        <li
-          key={locale}
-          onClick={() => {
-            if (locale !== currentLocale)
-              router.push({ pathname, query }, asPath, { locale }).finally(() => router.reload());
-          }}
-          data-active={locale === currentLocale}
-        >
-          {locale}
+        <li key={locale} data-active={locale === currentLocale}>
+          <Link href={basePath} locale={locale}>
+            {locale}
+          </Link>
         </li>
       ))}
     </ul>
