@@ -1,5 +1,5 @@
-import Link from 'next/link';
-import { useRouter } from 'next/router';
+import Link from 'next-intl/link';
+import { usePathname } from 'next/navigation';
 import { useState } from 'react';
 import { FcHome, FcReading } from 'react-icons/fc';
 import BurgerButton from './BurgerButton';
@@ -9,25 +9,25 @@ import ThemeSwitch from './ThemeSwitch';
 import styles from './Navigation.module.css';
 
 export default function Navigation() {
-  const router = useRouter();
+  const pathname = usePathname();
   const [active, setActive] = useState<boolean>(false);
 
   return (
     <nav className={styles.container} data-active={active}>
       <BurgerButton active={active} onClick={() => setActive(!active)} />
-      <div className={styles.title} data-active={router.pathname === '/' ? true : active}>
+      <div className={styles.title} data-active={pathname === '/' ? true : active}>
         PtCookie<span>.</span>DevLog
       </div>
       <div className={styles.content} data-active={active}>
         <ul>
           <li>
-            <Link href="/" className={router.pathname === '/' ? styles.link : undefined}>
+            <Link href="/" className={pathname === '/' ? styles.link : undefined}>
               home
             </Link>
             <FcHome className={styles.icon} />
           </li>
           <li>
-            <Link href="/posts" className={router.pathname.startsWith('/posts') ? styles.link : undefined}>
+            <Link href="/posts" className={pathname?.startsWith('/posts') ? styles.link : undefined}>
               posts
             </Link>
             <FcReading className={styles.icon} />
