@@ -1,20 +1,19 @@
+'use client';
+
+import Link from 'next-intl/link';
 import { ThemeProvider } from '@/components/context/ThemeProvider';
 import DefaultLayout from '@/components/layout/DefaultLayout';
 
-import type { NextPage } from 'next';
-
-type Props = {
-  locale?: string;
-  statusCode?: number;
-};
-
-const ErrorPage: NextPage<Props> = ({ locale, statusCode }) => {
+export default function NotFound() {
   return (
     <ThemeProvider>
       <DefaultLayout>
         <div>
-          {statusCode === 404 && <h1>404</h1>}
-          {locale === 'ko' ? <h1>페이지를 찾을 수 없습니다</h1> : <h1>Page not found</h1>}
+          <h1>404</h1>
+          <h2>Page not found</h2>
+          <p>
+            Back to <Link href="/">Home</Link>
+          </p>
         </div>
         <style jsx>{`
           div {
@@ -29,11 +28,4 @@ const ErrorPage: NextPage<Props> = ({ locale, statusCode }) => {
       </DefaultLayout>
     </ThemeProvider>
   );
-};
-
-ErrorPage.getInitialProps = ({ locale, res, err }) => {
-  const statusCode = res ? res.statusCode : err ? err.statusCode : 404;
-  return { locale, statusCode };
-};
-
-export default ErrorPage;
+}
