@@ -1,9 +1,5 @@
 import styles from './Copyright.module.css';
 
 export default function Copyright() {
-  return (
-    <>
-      <p className={styles.copyright}>&copy; 2022, PtCookie</p>
-    </>
-  );
+  return <p className={styles.copyright}>&copy; 2023, PtCookie</p>;
 }
