@@ -1,4 +1,4 @@
-import Link from 'next/link';
+import Link from 'next-intl/link';
 import { generatePagination } from '@/lib/pagination';
 
 import styles from './Pagination.module.css';
@@ -6,10 +6,7 @@ import styles from './Pagination.module.css';
 type Props = {
   current: number;
   pages: number;
-  link: {
-    href: (page: number) => string;
-    as: (page: number) => string;
-  };
+  link: (page: number) => string;
 };
 
 export default function Pagination({ current, pages, link }: Props) {
@@ -22,7 +19,7 @@ export default function Pagination({ current, pages, link }: Props) {
           {it.excerpt ? (
             '...'
           ) : (
-            <Link href={link.href(it.page)} as={link.as(it.page)} data-active={it.page === current}>
+            <Link href={link(it.page)} data-active={it.page === current}>
               {it.page}
             </Link>
           )}
