@@ -29,10 +29,7 @@ export default function TagPostList({ posts, tag, pagination }: Props) {
       <Pagination
         current={pagination.current}
         pages={pagination.pages}
-        link={{
-          href: () => '/posts/tags/[[...slug]]',
-          as: (page) => (page === 1 ? '/posts/tags/' + tag.slug : `/posts/tags/${tag.slug}/${page}`),
-        }}
+        link={(page) => (page === 1 ? `/posts/tags/${tag.slug}` : `/posts/tags/${tag.slug}/${page}`)}
       />
     </div>
   );
